backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "backend for Pyramid project"
authors = ["Izpodvypodvert <albukerke87@mail.ru>"]
license = "MIT"
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.12.1"
fastapi = "^0.110.0"
uvicorn = {extras = ["standart"], version = "^0.29.0"}
sqlmodel = "^0.0.16"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
setuptools = "^69.2.0"
fastapi-users-db-sqlmodel = "^0.3.0"
markdown = "^3.6"
bleach = "^6.1.0"
celery = "^5.3.6"
redis = "^5.0.3"
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
